Add Python script
The program runs so far but I'll need to do some error checking to make
sure the it gives the appropriate results. I'll need to download the
wrong file and make sure that it shows the file isn't authenticated.

It also does this weird behavior where the public key (the last file
downloaded) progress appears after the check. I'm going to have to learn
to submit a "Issue" on Github.