Add code to evaluate authenticity
There's been no checking to see if the program runs in the other cases
such as a compromised file or bogus input into the prompts.

There's also this weird behavior of showing the progress bar for the
public key AFTER the check has been performed. I'll have to learn how to
submit issues in Github.